+++ whatsnew.md
@@ -1,6 +1,7 @@
 #v1.3
 ----------------------
 
+* GUIObject have two more methods: _onHoverOn_ and _onHoverOut_ to add action corresponding to on pointer over and on pointer on
 * Method to update the layer mask of all existing meshes to make them invisible to the GUI camera
 * The text doesn't look crisp anymore !! YEAHHH !
 * Added a fade out animation for a gui object